First commit
"Swiftextual" will not be the actual name of this application. It is actually "Textual". I am just using this as a name while I work on rewriting Textual into Swift.

This first commit is the starting point.

Committing now a shared framework that will be accessed by all components of the application.